extract book navigation
